Added shotgun / "shutgun" functionality to the game. Added sprite on top of player-character when it is enabled.
You can press Q to enable shotgun (and later ALL weapons) at once, without requiring the items. This is for testing purposes.

The shotgun sprite is put on top of the player, by means of a UI-image.
PlayerItems now contain an enumerator that can be used to give any item the weapon-type desired.